requirements.txt
Flask==2.3.2
gunicorn==21.2.0
pandas
numpy
joblib
python-dotenv
requests
